Move criteria types into teacher tool
