+++ teachertool/src/state/state.ts
@@ -1,12 +1,13 @@
 import { ModalType, Notifications } from "../types";
+import { CatalogCriteria, CriteriaInstance } from "../types/criteria";
 
 export type AppState = {
     targetConfig?: pxt.TargetConfig;
     notifications: Notifications;
     currentEvalResult: pxt.blocks.EvaluationResult | undefined;
     projectMetadata: pxt.Cloud.JsonScript | undefined;
-    catalog: pxt.blocks.CatalogCriteria[] | undefined;
-    selectedCriteria: pxt.blocks.CriteriaInstance[];
+    catalog: CatalogCriteria[] | undefined;
+    selectedCriteria: CriteriaInstance[];
     modal: ModalType | undefined;
 };
